feat(seo): Optimize robots.txt for major search engine indexing
- Explicitly allow Googlebot, Bingbot, and Yahoo (Slurp)
- Block sensitive pages such as /admin/, /login/, /cart/, and /checkout/
- Add sitemap to improve indexing efficiency
- Maintain a clean structure following SEO best practices

This update aims to enhance site visibility on major search engines while preventing unnecessary page indexing.
